+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - _Created_ `enemyHealthBar.lua` component.
 - _Created_ `tooltip.lua` component.
 - Context-Menu entries now accept `Resolver Functions` that return dynamically generated values for context-menus during run-time.
+- _Created_ `ContextMenu:Add()` to handle table insertion properly. Can handle multiple entries at once.
 
 ### CHANGED